+++ src/main/native/include/ctre/phoenix/motorcontrol/GroupMotorControllers.h
@@ -10,7 +10,7 @@ namespace motorcontrol {
 class GroupMotorControllers {
 public:
 	static void Register(IMotorController *motorController);
-	static int MotorControllerCount();
+	static size_t MotorControllerCount();
 	static IMotorController* Get(int idx);
 
 private:
